Fix issues introduced with autocasting game update (#21)
A polish to the game's autocasting system released today, 2025-04-09.
Full details are available
[here](https://secure.runescape.com/m=news/a=13/farming--autocast-qol-improvements?oldschool=1),
but I've enumerated the changes relevant to this plugin:

- Reduced magic level / running out of runes, and performing a combat
action afterwards (typically a melee) no longer unsets the current
autocast varbit.
- An autocast can be set without having the proper runes / level
requirements.
- Varbit remains even when changing to a different weapon which can
autocast, but not on the current spellbook.
- Varbit remains even between sessions, meaning the plugin should read
all this on login (first login on the client especially) and correctly
interpret the state. Game data is loaded in stages, e.g. the autocasting
varbit is set first, before inventory and equipment is known. The plugin
currently will hit a null pointer and crash.

Other than the login issue, which is fixed by adding another null check,
most of these are fixed by removing assumptions. For example, if the
autocasting varbit is set and we read an equipment change event, we were
assuming then you had the level and runes to cast and set flags
accordingly. We can no longer do this; state updates need to happen only
by the relevant events.

All that to say, these were resolvable with two fixes.

I also updated the build.gradle because mine was out of date and not
reading imports correctly. New version is directly from the plugin hub's
latest template.